Константы порта и smtp ушли в setting.py
