+++ src/core/settings.py
@@ -11,12 +11,10 @@ TELEGRAM_TOKEN = os.getenv('TELEGRAM_TOKEN')
 BASE_DIR = Path(__file__).resolve().parent.parent
 LOG_LEVEL = 'INFO'
 
+
+# Настройки для отправки email-сообщения куратору
+PORT_SMTP_SERVER = 587
+SMTP_SERVER = 'smtp.yandex.ru'
 EMAIL_BOT = os.getenv('EMAIL_BOT', 'FondProvidenieBot@yandex.ru')
-
-# ЗАМЕНИТЬ НА МЫЛО КУРАТОРА ФОНДА
-EMAIL_CURATOR = os.getenv('EMAIL_CURATOR', 'k.danilow2009@yandex.ru')
+EMAIL_CURATOR = os.getenv('EMAIL_CURATOR', 'k.danilow2009@yandex.ru') # ЗАМЕНИТЬ НА МЫЛО КУРАТОРА ФОНДА
 EMAIL_BOT_PASSWORD = os.getenv('EMAIL_BOT_PASSWORD', 'jyvsejdjxyixsxkh')
-
-TELEGRAM_TOKEN = os.getenv("TELEGRAM_TOKEN")
-BASE_DIR = Path(__file__).resolve().parent.parent
-LOG_LEVEL = "INFO"
